Fix syntax table for triple quotes
Like `python-mode`, `graphql-mode` has the ability to set block strings
using triple-quoted blocks. By default, these triple-quoted strings end
up messing up if you place a single double-quote within them.

This change modifies the default settings for propertizing characters in
the syntax table. For triple-quoted blocks, we now set the string fence
syntax on the last quote of the opening fence and the first quote of the
closing fence.

This allows for normal `forward-sexp` and `backward-sexp` movements in
`graphql-mode` more readily than in `python-mode` because of the
conventions around block strings in the GraphQL spec. For example:

```
"""
This is a good block string.
"""

"""Whereas this is not."""
```

I lifted the vast majority of this change directly from `python-mode`.